Further refactor + remove dead code
