--- src/components/ItemList.vue
@@ -1,130 +1,0 @@
-<template>
-  <div class="news-view">
-    <spinner :show="loading"></spinner>
-
-    <div class="news-list-nav">
-      <router-link v-if="page > 1" :to="'/' + type + '/' + (page - 1)">&lt; prev</router-link>
-      <a v-else class="disabled">&lt; prev</a>
-      <span>{{ page }}/{{ maxPage }}</span>
-      <router-link v-if="hasMore" :to="'/' + type + '/' + (page + 1)">more &gt;</router-link>
-      <a v-else class="disabled">more &gt;</a>
-    </div>
-
-    <transition :name="transition">
-      <div class="news-list" :key="displayedPage" v-if="displayedPage > 0">
-        <transition-group tag="ul" name="item">
-          <item v-for="item in displayedItems" :key="item._id" :item="item">
-          </item>
-        </transition-group>
-      </div>
-    </transition>
-  </div>
-</template>
-
-<script>
-import Spinner from './Spinner.vue'
-import Item from './Item.vue'
-
-let isInitialRender = true
-
-export default {
-  name: 'item-list',
-
-  components: {
-    Spinner,
-    Item
-  },
-
-  props: {
-    type: String
-  },
-
-  data () {
-    const data = {
-      loading: false,
-      transition: 'slide-up',
-      // if this is the initial render, directly render with the store state
-      // otherwise this is a page switch, start with blank and wait for data load.
-      // we need these local state so that we can precisely control the timing
-      // of the transitions.
-      displayedPage: isInitialRender ? Number(this.$store.state.route.params.page) || 1 : -1,
-      displayedItems: [{
-        id: 1,
-        score: 1,
-        title: 'Testing'
-      }]
-      // displayedItems: isInitialRender ? this.$store.getters.activeItems : []
-    }
-    isInitialRender = false
-    return data
-  },
-
-  computed: {
-    page () {
-      return Number(this.$store.state.route.params.page) || 1
-    },
-    maxPage () {
-      const { itemsPerPage, lists } = this.$store.state
-      return Math.ceil(lists[this.type].length / itemsPerPage)
-    },
-    hasMore () {
-      return this.page < this.maxPage
-    }
-  }
-}
-</script>
-
-<style lang="stylus">
-
-.news-list-nav, .news-list
-  background-color #fff
-  border-radius 2px
-
-.news-list-nav
-  padding 15px 30px
-  position fixed
-  text-align center
-  top 55px
-  left 0
-  right 0
-  z-index 998
-  box-shadow 0 1px 2px rgba(0,0,0,.1)
-  a
-    margin 0 1em
-  .disabled
-    color #ccc
-
-.news-list
-  position absolute
-  margin 30px 0
-  width 100%
-  transition all .5s cubic-bezier(.55,0,.1,1)
-  ul
-    list-style-type none
-    padding 0
-    margin 0
-
-.slide-left-enter, .slide-right-leave-active
-  opacity 0
-  transform translate(30px, 0)
-
-.slide-left-leave-active, .slide-right-enter
-  opacity 0
-  transform translate(-30px, 0)
-
-.item-move, .item-enter-active, .item-leave-active
-  transition all .5s cubic-bezier(.55,0,.1,1)
-
-.item-enter
-  opacity 0
-  transform translate(30px, 0)
-
-.item-leave-active
-  position absolute
-  opacity 0
-  transform translate(30px, 0)
-
-@media (max-width 600px)
-  .news-list
-    margin 10px 0
-</style>
